requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# HTTP client
requests==2.31.0

# Data science and ML
pandas>=2.2.0
scikit-learn==1.4.0
numpy>=1.26.0
scipy>=1.12
statsmodels==0.14.1

# Deep learning frameworks
tensorflow>=2.16.0
torch>=2.2.0
transformers>=4.36.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0
joblib==1.3.2

# Async and networking
websockets==12.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional GPU acceleration
# cupy-cuda12x==12.2.0  # Uncomment for CUDA support
# tensorflow-gpu==2.16.0  # Uncomment for GPU TensorFlow
